Split feed info and settings
