+++ core/src/main/res/values/strings.xml
@@ -122,6 +122,9 @@
     <string name="mark_all_seen_msg">Marked all Episodes as seen</string>
     <string name="mark_all_seen_confirmation_msg">Please confirm that you want to mark all episodes as seen.</string>
     <string name="show_info_label">Show information</string>
+    <string name="show_feed_settings_label">Show feed settings</string>
+    <string name="feed_info_label">Feed info</string>
+    <string name="feed_settings_label">Feed settings</string>
     <string name="rename_feed_label">Rename Podcast</string>
     <string name="remove_feed_label">Remove Podcast</string>
     <string name="share_label">Share&#8230;</string>
